#175 refactor websocket message handling to redux
